// cbpv-platform/apps/web/src/gen/api/task/model/answer.ts
/**
 * Task
 * Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Answer { 
    id: string;
    textAnswer: string;
    radioAnswer: string;
    starRatingAnswer: number;
    thumbsRatingAnswer: Answer.ThumbsRatingAnswerEnum;
    feedbackId: string;
    questionId: string;
}
export namespace Answer {
    export type ThumbsRatingAnswerEnum = 'up' | 'down';
    export const ThumbsRatingAnswerEnum = {
        Up: 'up' as ThumbsRatingAnswerEnum,
        Down: 'down' as ThumbsRatingAnswerEnum
    };
}