// cbpv-platform/apps/web/src/gen/api/task/model/feedback.ts
/**
 * Task
 * Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Answer } from './answer';
import { Task } from './task';


export interface Feedback { 
    id: string;
    dateCreated: string;
    comment: string;
    paymentStatus: Feedback.PaymentStatusEnum;
    feedbackRating: number;
    feedbackRatingComment: string;
    taskRating: number;
    taskRatingComment: string;
    taskId: string;
    task: Task;
    userId: string;
    answers: Array<Answer>;
}
export namespace Feedback {
    export type PaymentStatusEnum = 'pending' | 'completed';
    export const PaymentStatusEnum = {
        Pending: 'pending' as PaymentStatusEnum,
        Completed: 'completed' as PaymentStatusEnum
    };
}