// cbpv-platform/apps/web/src/gen/api/task/model/task.ts
/**
 * Task
 * Task service is responsible for iterative validation of Prototypes. Tasks are created and Feedback is provided through this service.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Skill } from './skill';
import { Feedback } from './feedback';
import { Country } from './country';
import { Question } from './question';
import { Image } from './image';


export interface Task { 
    id: string;
    title: string;
    description: string;
    dateCreated: string;
    testType: Task.TestTypeEnum;
    prototypeFormat: Task.PrototypeFormatEnum;
    status: Task.StatusEnum;
    accessType: Task.AccessTypeEnum;
    iframeUrl1: string;
    iframeUrl2: string;
    textualDescription1: string;
    textualDescription2: string;
    minAge: number;
    maxAge: number;
    minExperience: number;
    maxExperience: number;
    budget: number;
    incentive: number;
    userId: string;
    projectId: string;
    images: Array<Image>;
    skills: Array<Skill>;
    countries: Array<Country>;
    questions: Array<Question>;
    feedbacks: Array<Feedback>;
}
export namespace Task {
    export type TestTypeEnum = 'basic' | 'comparison';
    export const TestTypeEnum = {
        Basic: 'basic' as TestTypeEnum,
        Comparison: 'comparison' as TestTypeEnum
    };
    export type PrototypeFormatEnum = 'image' | 'iframe' | 'text';
    export const PrototypeFormatEnum = {
        Image: 'image' as PrototypeFormatEnum,
        Iframe: 'iframe' as PrototypeFormatEnum,
        Text: 'text' as PrototypeFormatEnum
    };
    export type StatusEnum = 'draft' | 'open' | 'closed';
    export const StatusEnum = {
        Draft: 'draft' as StatusEnum,
        Open: 'open' as StatusEnum,
        Closed: 'closed' as StatusEnum
    };
    export type AccessTypeEnum = 'open' | 'nda' | 'request';
    export const AccessTypeEnum = {
        Open: 'open' as AccessTypeEnum,
        Nda: 'nda' as AccessTypeEnum,
        Request: 'request' as AccessTypeEnum
    };
}